refactor: lift helpers out of useBaseplateGeneration and the fit suggestion out of useDrawerSettings
useBaseplateGeneration's module-level helpers move to generationMeshEntries
and generationTriggers unchanged. useDrawerSettings hands the measured-drawer
fit suggestion (its anchored state, dismissal memory and four callbacks) to a
new useDrawerFitSuggestion hook that takes the drawer values as inputs and
returns the same names into the same return object.
